pyproject.toml
[project]
name = "kabi-discord-cli"
version = "0.2.8"
description = "Discord CLI for local-first sync, search, export, and agent-friendly retrieval"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "jackwener" }]
keywords = ["discord", "cli", "sqlite", "agent", "discord-cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Communications :: Chat",
    "Topic :: Utilities",
]
urls = { Homepage = "https://github.com/jackwener/discord-cli", Repository = "https://github.com/jackwener/discord-cli", Issues = "https://github.com/jackwener/discord-cli/issues" }
dependencies = [
    "httpx>=0.27",
    "click>=8.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["build>=1.2", "pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.11", "twine>=6.0"]


[project.scripts]
discord = "discord_cli.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "src/discord_cli",
    "tests",
    "README.md",
    "README_CN.md",
    "SKILL.md",
    ".env.example",
    "LICENSE",
    "CHANGELOG.md",
    "CONTRIBUTING.md",
]

[tool.ruff]
target-version = "py310"
line-length = 100
